Simplify component state binding interface
This change increases the specificity of the API, more tightly coupling
Coherence with React, in exhange for a more convenient API. This is
appropriate, given Coherence is specifically for managing React
component state.

Replace `data` and `addChangeListener` with a `store.subscribe` method,
that accepts the component, and named subjects.

Define store data in terms of RxJS BehaviorSubject, rather than managing
a state object directly. Use the Subject API to add subscribers, and to
push events to subscribers, rather than using EventEmitter.

Add React integration tests.