Promote React Component-bindable state as an independent Model class
Defining behavior in terms of Actions is the responsibility of
Controllers (formerly, Coherence stores).

Previously, Coherence stores also were responsible for exposing state
that could change as a result of Actions, to React components. However,
this was problematic, because in some cases, Coherence stores needed to
share state. This was difficult, if that state was encapsulated within
the store.

The solution is to reduce the responsibility of stores, and to define
and declare state independently, as Models. Stores become Controllers,
that perform operations on Models. Any Controller can operate on any
Model.

The Model binds to the Component directly, rather than through a Store.
This way, any changes to the Model's state, regardless of the origin,
can be propagated to the bound React components.