Replace controllers with Intents
The "Controller" pattern is an arbitrary abstraction layer, around
responding to actions.

Dispatcher registration, and switching on an action "type" string is
error-prone.

Direct subscription to an Intent class is safer since one can't
subscribe to a non-existent Intent. They'll get a "no method subscribe
on undefined", whereas one might try to respond to a non-existing action
type via typo, and be non the wiser.

Routing is removed. The intention is to factor this out into an optional
companion module, which implements special Intents, with baked-in
routing, and optional pushState hooks.